move deadline-related stuff to util/workspace.ts
